requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
redis>=6.0.0
python-decouple>=3.8
djangorestframework-simplejwt>=5.3.0
msal>=1.25.0
whitenoise>=6.6.0
celery>=5.3.0
Pillow>=10.0.0
django-filter>=23.0
pytest-django>=4.5.0
factory-boy>=3.3.0
coverage>=7.0.0
gunicorn>=21.0.0